feat: byte-parity with reference assemblers, complete instruction set, working AmigaDOS extraction
Verified byte-for-byte against a reference assembler on two independent
corpora of real Amiga sources — 25/25 files now assemble identically:

  * 12/12 in a demo/game codebase (up to 524 KB)
  * 13/13 in a tutorial repo built on the original Commodore system
    includes (exec/, graphics/, intuition/, ...)

Three further files are rejected by the reference assembler itself; on one
we are more permissive (BRA.S with displacement 0, where the short form is
reserved, widens to the word form instead of erroring out).

Instruction set (25 fixes, verified against Kickstart ROMs)
-----------------------------------------------------------
Scc pattern, 13 missing decoder patterns (MULU.L/DIVSL.L/CAS/CHK2/TRAPcc/
RTM/CALLM/MOVES/MOVE16/PSAVE/PRESTORE/EXTB.L/PACK/UNPK/LPSTOP), MOVEA,
MOVES, ABCD/SBCD, ADDX/SUBX/CMPM, ADDQ/SUBQ on Dn, MOVE extension words,
ORI/ANDI/EORI-to-CCR/SR, MOVE SR/CCR/USP, index address registers, negative
hex displacements, PC-relative displacements. Covered by a new roundtrip
test over ~280 instruction forms (assemble -> decode -> reassemble).

ROM roundtrip rate (per instruction, 68020): KS1.3 82.4%, KS3.2.3 79.3%
(was 64.1%/63.6%).

Assembler syntax and semantics
------------------------------
* INCLUDE was never wired up (handle_include existed but was never
  called). Now expanded before macro processing, include-once, with a new
  -I/--include search path.
* Local labels (.loop) were missing entirely — scoped per global label,
  tracked identically across both passes and branch relaxation.
* EQU was overwritten with the PC by recalculate_pcs, so every constant
  resolved to its own address.
* Expressions as displacements (DMACON-CUSTOM_BASE(A6)) evaluated to 0.
* Labels without a trailing colon were not recognised.
* `*` as a comment marker was unsupported, while keeping its meaning as
  multiply and as the PC symbol.
* IFD/IFND/ENDC (the Motorola spellings) were missing — every Amiga
  header's include guard.
* Macro arguments the invocation omitted stayed literal `\1` instead of
  substituting empty, breaking the standard IFC '\1','' idiom.
* CMP/EOR/ADD/SUB/AND/OR with an immediate and a memory destination were
  rejected instead of selecting CMPI/EORI/ADDI/...

Encoding conformance
--------------------
* Absolute addresses and branches were always shortened; reference
  assemblers only do so when optimizing. New --optimize flag, off by
  default. An explicit .S/.W/.L suffix is now binding.
* DC.B rounded every directive up to an even length (05 00 EE 00 instead
  of 05 EE); instructions now get a pad byte instead.
* EVEN emitted a 2-byte NOP word rather than one zero byte; CNOP
  overshot on an odd PC.
* A trailing DS/DCB was dropped from binary output.
* (d,PC,Xn) was two bytes off, labels in both PC-relative forms resolved
  to 0, and EAOperand::PcIndex had its target/displacement fields swapped.

m68k-floppy: AmigaDOS against real Workbench disks
--------------------------------------------------
Both bugs were invisible because reader and writer shared the same wrong
assumption, so synthetic images round-tripped cleanly:

* The file header's data-block table is anchored at the end of the block
  and grows downwards. The reader walked it from index 0 — all zeroes on
  a real disk, so every extraction produced an empty file.
* amiga_hash masked to 31 bits instead of 11 (0x7FF), naming the wrong
  bucket for all but the shortest names (~30% hit rate, i.e. chance), so
  --extract reported "not found" for most files.

Verified: 981 files across all eight Workbench 1.3/3.1 disks extract at
their exact advertised length, magic numbers spot-checked.

New in amigados.rs: resolve_path/read_file_at_path/list_dir_at_path/walk.
CLI gains --list [PATH], --list-all, --extract PATH, --extract-all DIR
(with path-traversal guarding). Previously only the root directory was
reachable — 17 of 170 entries on a Workbench 3.1 disk.

Tests: 588 passing (was 570), clippy and fmt clean.

+++ crates/m68k-core/src/operands.rs
@@ -24,9 +24,14 @@ pub enum Operand {
     AbsoluteLong(i32),
     /// Program counter with displacement: (d16,PC) or (d32,PC)
     PcRelativeDisp(i32, bool),
-    /// Program counter with index: (d8,PC,Xn*scale). Fields: Xn, disp,
-    /// scale, index size (`true` = `.L`, `false` = `.W`).
-    PcRelativeIndex(u8, i8, u8, bool),
+    /// Program counter with index: (d8,PC,Xn*scale). Fields: Xn, target
+    /// address, scale, index size (`true` = `.L`, `false` = `.W`).
+    ///
+    /// Like [`Operand::PcRelativeDisp`], the second field holds the full
+    /// *target address*, not the encoded displacement — the parser cannot
+    /// subtract the PC because it only sees the operand text. The encoder
+    /// narrows it against the extension word's address.
+    PcRelativeIndex(u8, i32, u8, bool),
     /// Immediate: #expr
     Immediate(i64),
     /// Memory reference (general expression)
